feat: Move facts and figures to sections
