+++ src/app/zulieferer/page.tsx
@@ -3,7 +3,7 @@
 import Hero from "./Hero";
 import SustainableProcurement from "./SustainableProcurement";
 import FAQ from "../../components/sections/FAQ";
-import FactsAndFigures from "@/components/FactsAndFigures";
+import FactsAndFigures from "@/components/sections/FactsAndFigures";
 import Quote from "@/components/sections/Quote";
 import Contact from "@/components/sections/Contact";
 import TeaserDouble from "@/components/sections/TeaserDouble";
